renaming package to kgcl-rdflib
