+++ kgcl_rdflib/diff/diff_2_kgcl_existential.py
@@ -2,9 +2,9 @@
 
 import rdflib
 
-from kgcl.diff.graph_diff import (get_added_existentials,
-                                  get_deleted_existentials)
-from kgcl.diff.render_operations import render
+from kgcl_rdflib.diff.graph_diff import (get_added_existentials,
+                                         get_deleted_existentials)
+from kgcl_rdflib.diff.render_operations import render
 from kgcl_schema.datamodel.kgcl import EdgeCreation, EdgeDeletion
