+++ kgcl_rdflib/diff/diff_2_kgcl_triple_annotation.py
@@ -2,9 +2,9 @@
 import rdflib
 from rdflib import Graph, Literal, URIRef
 
-from kgcl.diff.graph_diff import (get_added_triple_annotations,
-                                  get_deleted_triple_annotations)
-from kgcl.diff.render_operations import render
+from kgcl_rdflib.diff.graph_diff import (get_added_triple_annotations,
+                                         get_deleted_triple_annotations)
+from kgcl_rdflib.diff.render_operations import render
 from kgcl_schema.datamodel.kgcl import EdgeCreation, EdgeDeletion
 from kgcl_schema.datamodel.ontology_model import Annotation
