+++ kgcl_rdflib/diff/render_operations.py
