+++ kgcl_rdflib/diff/summary_generation.py
@@ -5,9 +5,9 @@ from datetime import datetime
 import rdflib
 from rdflib.util import guess_format
 
-import kgcl.diff.diff_2_kgcl_existential as existential
-import kgcl.diff.diff_2_kgcl_single as single
-from kgcl.diff.pretty_print_kgcl import render_instances
+import kgcl_rdflib.diff.diff_2_kgcl_existential as existential
+import kgcl_rdflib.diff.diff_2_kgcl_single as single
+from kgcl_rdflib.diff.pretty_print_kgcl import render_instances
 
 
 def ts():
