+++ kgcl_rdflib/kgcl.py
@@ -6,7 +6,7 @@ import click
 import rdflib
 from rdflib.util import guess_format
 
-from kgcl.apply import graph_transformer
+from kgcl_rdflib.apply import graph_transformer
 from kgcl_schema.grammar import parser
 
 # TODO: remove this
