+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "kgcl"
+name = "kgcl-rdflib"
 version = "0.1.0"
 description = "Schema fro the KGCL project."
 authors = [
