+++ tests/test_cli/test_cli.py
@@ -4,8 +4,8 @@ import unittest
 import rdflib
 from click.testing import CliRunner
 
-import kgcl.kgcl as kgcl_apply
-import kgcl.kgcl_diff as kgcl_diff
+import kgcl_rdflib.kgcl as kgcl_apply
+import kgcl_rdflib.kgcl_diff as kgcl_diff
 import kgcl_schema.grammar.parser as kgcl_parser
 from kgcl_schema.datamodel.kgcl import Session
 from kgcl_schema.utils import from_yaml
