+++ tests/util.py
@@ -5,7 +5,7 @@ from linkml_runtime.utils.yamlutils import YAMLRoot
 from rdflib import Graph
 from rdflib.compare import graph_diff, to_isomorphic
 
-from kgcl.apply.graph_transformer import apply_patch
+from kgcl_rdflib.apply.graph_transformer import apply_patch
 from kgcl_schema.grammar.parser import parse
